task10/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
PyJWT
